# backend-go/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Devaulty API Documentation
  description: REST API for Devaulty Desktop App Engine
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development Server

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: DEVAULTY_INTERNAL_TOKEN
      description: |
        Internal security token generated in memory per app lifecycle.
        Note: When running in DEV environment (APP_ENV=dev), this token defaults to 'dev-token' for interactive testing.

  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid project ID"

    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Project archived"

    Project:
      type: object
      required:
        - id
        - name
        - archived
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "06f891ae-368b-4bd6-b1e6-d957f1496f9a"
        name:
          type: string
          example: "Devaulty Go"
        description:
          type: string
          nullable: true
          example: "Developer Vault Application"
        icon:
          type: string
          nullable: true
          example: "rocket"
        color:
          type: string
          nullable: true
          example: "#4A90E2"
        archived:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2026-08-01T12:35:07.84097684-03:00"
        updatedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-08-01T16:22:46.12345678-03:00"

    CreateProjectCommand:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
          example: "Devaulty Go"
        description:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
          example: "Developer Vault Application"
        icon:
          type: string
          maxLength: 100
          nullable: true
          example: "rocket"
        color:
          type: string
          nullable: true
          pattern: "^#([A-Fa-f0-9]{3,4}|[A-Fa-f0-9]{6}|[A-Fa-f0-9]{8})$"
          example: "#4A90E2"

    UpdateProjectCommand:
      type: object
      properties:
        name:
          type: string
          nullable: true
          example: "Updated Devaulty Name"
        description:
          type: string
          nullable: true
          example: "Updated Description"
        icon:
          type: string
          nullable: true
          example: "folder"
        color:
          type: string
          nullable: true
          example: "#00FF00"

    ProjectPage:
      type: object
      required:
        - content
        - size
        - number
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Project'
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        totalElements:
          type: integer
          format: int64
          example: 1
        totalPages:
          type: integer
          example: 1

    SnippetType:
      type: string
      enum:
        - COMMAND
        - CODE
      example: "CODE"

    SnippetLanguage:
      type: string
      nullable: true
      enum:
        - BASH
        - FISH
        - ZSH
        - SH
        - POWERSHELL
        - BATCH
        - JAVA
        - KOTLIN
        - JAVASCRIPT
        - TYPESCRIPT
        - PYTHON
        - GO
        - RUST
        - C
        - CPP
        - CSHARP
        - PHP
        - RUBY
        - SWIFT
        - DART
        - SCALA
        - LUA
        - PERL
        - R
        - ELIXIR
        - HASKELL
        - CLOJURE
        - GROOVY
        - HTML
        - CSS
        - SCSS
        - LESS
        - JSX
        - TSX
        - VUE
        - SVELTE
        - JSON
        - YAML
        - XML
        - TOML
        - INI
        - ENV
        - CSV
        - MARKDOWN
        - PROPERTIES
        - DOCKERFILE
        - DOCKER_COMPOSE
        - NGINX
        - APACHE
        - TERRAFORM
        - ANSIBLE
        - KUBERNETES_YAML
        - HELM
        - MAKEFILE
        - CMAKE
        - GRADLE
        - MAVEN_POM
        - SQL
        - PLSQL
        - GRAPHQL
        - MONGODB
        - GITHUB_ACTIONS
        - GITLAB_CI
        - JENKINSFILE
        - REGEX
        - DIFF
        - LOG
        - PLAIN_TEXT
      example: "GO"

    Snippet:
      type: object
      required:
        - id
        - projectId
        - title
        - content
        - snippetType
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "784f6f5c-57af-4313-b160-688a3e0d283b"
        projectId:
          type: string
          format: uuid
          example: "06f891ae-368b-4bd6-b1e6-d957f1496f9a"
        title:
          type: string
          example: "Print Hello World"
        description:
          type: string
          nullable: true
          example: "Simple print example in Go"
        content:
          type: string
          example: "fmt.Println(\"Hello World\")"
        language:
          $ref: '#/components/schemas/SnippetLanguage'
        snippetType:
          $ref: '#/components/schemas/SnippetType'
        createdAt:
          type: string
          format: date-time
          example: "2026-08-03T12:35:07.84097684-03:00"
        updatedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-08-03T16:22:46.12345678-03:00"

    CreateSnippetCommand:
      type: object
      required:
        - title
        - content
        - language
        - snippetType
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 255
          example: "Print Hello World"
        description:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
          example: "Simple print example in Go"
        content:
          type: string
          minLength: 1
          example: "fmt.Println(\"Hello World\")"
        language:
          $ref: '#/components/schemas/SnippetLanguage'
        snippetType:
          $ref: '#/components/schemas/SnippetType'

    UpdateSnippetCommand:
      type: object
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 255
          nullable: true
          example: "Updated Snippet Title"
        description:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
          example: "Updated description"
        content:
          type: string
          minLength: 1
          nullable: true
          example: "log.Println(\"Updated Content\")"
        language:
          $ref: '#/components/schemas/SnippetLanguage'
        snippetType:
          $ref: '#/components/schemas/SnippetType'

    SnippetPage:
      type: object
      required:
        - content
        - size
        - number
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/Snippet'
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
        totalElements:
          type: integer
          format: int64
          example: 1
        totalPages:
          type: integer
          example: 1

security:
  - ApiKeyAuth: []

paths:
  /projects:
    post:
      summary: Create a new project
      description: Creates a new project entity in the database.
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectCommand'
      responses:
        '201':
          description: Project created successfully
          headers:
            Location:
              schema:
                type: string
              description: Relative URL path to the newly created project
              example: "/api/v1/projects/06f891ae-368b-4bd6-b1e6-d957f1496f9a"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid JSON format or validation constraints failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get paginated list of projects
      description: Fetches a paginated list of projects ordered by creation date descending.
      tags:
        - Projects
      parameters:
        - name: page
          in: query
          description: Zero-based page index
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Page size limit
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated list of projects retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /projects/{id}:
    get:
      summary: Get project by ID
      description: Retrieves a single project entity by its UUID.
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          description: Project UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid project ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update project (Partial Update)
      description: Performs a partial update (PATCH) on project fields specified in the request body.
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          description: Project UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectCommand'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Invalid UUID format or invalid JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete project by ID
      description: Permanently deletes a project by its UUID.
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          description: Project UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Project deleted successfully (No Content)
        '400':
          description: Invalid UUID format or delete error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /projects/{id}/archive:
    patch:
      summary: Archive project
      description: Sets project archived status to true.
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          description: Project UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid UUID or project already archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /projects/{id}/unarchive:
    patch:
      summary: Unarchive project
      description: Sets project archived status back to false.
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          description: Project UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project unarchived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid UUID or project already active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /projects/{project_id}/snippets:
    post:
      summary: Create a new snippet in a project
      description: Creates a new code or command snippet linked to the specified project.
      tags:
        - Snippets
      parameters:
        - name: project_id
          in: path
          description: Parent Project UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSnippetCommand'
      responses:
        '201':
          description: Snippet created successfully
          headers:
            Location:
              schema:
                type: string
              description: Relative URL path to the newly created snippet
              example: "/api/v1/projects/06f891ae-368b-4bd6-b1e6-d957f1496f9a/snippets/784f6f5c-57af-4313-b160-688a3e0d283b"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '400':
          description: Invalid JSON format or validation constraints failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get paginated list of snippets for a project
      description: Fetches a paginated list of snippets for the specified project ordered by creation date descending.
      tags:
        - Snippets
      parameters:
        - name: project_id
          in: path
          description: Parent Project UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Zero-based page index
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Page size limit
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated list of snippets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnippetPage'
        '400':
          description: Invalid query parameters or UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /projects/{project_id}/snippets/{snippet_id}:
    get:
      summary: Get snippet by ID
      description: Retrieves a single snippet entity by its UUID and parent project UUID.
      tags:
        - Snippets
      parameters:
        - name: project_id
          in: path
          description: Parent Project UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: snippet_id
          in: path
          description: Snippet UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Snippet details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '400':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or Snippet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update snippet (Partial Update)
      description: Performs a partial update (PATCH) on snippet fields specified in the request body.
      tags:
        - Snippets
      parameters:
        - name: project_id
          in: path
          description: Parent Project UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: snippet_id
          in: path
          description: Snippet UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSnippetCommand'
      responses:
        '200':
          description: Snippet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '400':
          description: Invalid UUID format or invalid JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or Snippet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete snippet by ID
      description: Permanently deletes a snippet by its UUID within a project.
      tags:
        - Snippets
      parameters:
        - name: project_id
          in: path
          description: Parent Project UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: snippet_id
          in: path
          description: Snippet UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Snippet deleted successfully (No Content)
        '400':
          description: Invalid UUID format or delete error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid DEVAULTY_INTERNAL_TOKEN header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or Snippet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'